+++ Assigment3/app/controllers/admin/orders_controller.rb
@@ -2,7 +2,7 @@
 
 module Admin
   class OrdersController < Admin::BaseController
-    layout 'admin_dashboard'
+    layout 'order_section'
     def index
       @orders = Order.all.includes(:user, order_items: :product).order(created_at: :desc)
       render 'orders/index'
