+++ Assigment3/app/models/product.rb
@@ -9,4 +9,5 @@ class Product < ApplicationRecord
 
   has_many :cart_items, dependent: :destroy, class_name: 'Cart'
   has_many :users, dependent: :destroy, through: :cart_items
+  has_many :order_items, dependent: :destroy
 end
