+++ Assigment3/app/views/orders/_user_orders.html.erb
@@ -1,0 +1,50 @@
+<div class="mb-10 p-6 rounded-lg shadow-xl backdrop-blur-sm bg-white/70 border border-gray-200">
+  <h2 class="text-2xl font-bold mb-4 text-gray-800">
+    <%= user.username %> (<%= user.email %>)
+  </h2>
+
+  <% orders.each do |order| %>
+    <div class="mt-6 pb-4 border-b border-gray-300">
+      <div class="flex flex-col md:flex-row md:justify-between md:items-center mb-4">
+        <span class="text-gray-700 text-sm md:text-base mb-2 md:mb-0">
+        <% if current_user&.admin? %>
+          <strong>Order ID:</strong> <%= order.id %> |
+        <% end %>
+          <strong>Status:</strong> <%= order.status.titleize %> |
+          <strong>Total:</strong> $<%= order.total_amount %> |
+          <strong>Day:</strong> <%= order.created_at.strftime("%Y-%m-%d") %> |
+          <strong>Hour:</strong> <%= order.created_at.strftime("%H:%M") %>
+        </span>
+
+        <% if current_user&.admin? && !%w[completed cancelled].include?(order.status) %>
+          <%= form_with(model: [:admin, order], method: :patch, local: true,
+                html: { class: "flex items-center space-x-2 mt-2 md:mt-0" }) do |form| %>
+            <%= form.select :status, Order.statuses.keys.map { |s| [s.titleize, s] },
+                { selected: order.status },
+                class: "border-2 border-gray-300 rounded-md px-3 py-1 bg-white text-gray-800 focus:outline-none focus:ring-2 focus:ring-orange-500" %>
+            <%= form.submit "Update", class: "px-4 py-1.5 bg-orange-600 text-white font-semibold rounded-md shadow-md hover:bg-orange-700 transition-colors cursor-pointer" %>
+          <% end %>
+        <% end %>
+      </div>
+
+      <table class="mt-4 w-full border-collapse">
+        <thead class="bg-gray-200/80">
+          <tr>
+            <th class="p-3 text-left text-gray-700 font-semibold rounded-tl-md">Product</th>
+            <th class="p-3 text-center text-gray-700 font-semibold">Quantity</th>
+            <th class="p-3 text-center text-gray-700 font-semibold rounded-tr-md">Price at Purchase</th>
+          </tr>
+        </thead>
+        <tbody>
+          <% order.order_items.each do |item| %>
+            <tr class="border-t border-gray-200 hover:bg-gray-50/50">
+              <td class="p-3 text-gray-800"><%= item.product.name %></td>
+              <td class="p-3 text-center text-gray-600"><%= item.quantity %></td>
+              <td class="p-3 text-center text-gray-800">$<%= item.price_at_purchase %></td>
+            </tr>
+          <% end %>
+        </tbody>
+      </table>
+    </div>
+  <% end %>
+</div>
