+++ Assigment3/app/views/orders/index.html.erb
@@ -1,58 +1,18 @@
 <% content_for :styles do %>
-    <%= stylesheet_link_tag "tailwind", "data-turbo-track": "reload", media: "all" %>
+  <%= stylesheet_link_tag "tailwind", "data-turbo-track": "reload", media: "all" %>
 <% end %>
 
 <div class="container mx-auto p-4 sm:p-8">
-    <h1 class="text-3xl text-white font-extrabold mb-8 text-center drop-shadow-lg">Orders Overview</h1>
+    <%= link_to "Back", root_path, class: "bg-red-500 text-white px-4 py-2 rounded hover:bg-red-600" %>
+    <h1 class="text-3xl text-white font-extrabold mb-8 text-center drop-shadow-lg">
+    Orders Overview
+    </h1>
 
+    <% if current_user&.admin? %>
     <% @orders.group_by(&:user).each do |user, orders| %>
-    <div class="mb-10 p-6 rounded-lg shadow-xl backdrop-blur-sm bg-white/70 border border-gray-200">
-        <h2 class="text-2xl font-bold mb-4 text-gray-800">
-            <%= user.username %> (<%= user.email %>)
-        </h2>
-
-        <% orders.each do |order| %>
-        <div class="mt-6 pb-4 border-b border-gray-300">
-            <div class="flex flex-col md:flex-row md:justify-between md:items-center mb-4">
-                <span class="text-gray-700 text-sm md:text-base mb-2 md:mb-0">
-                    <strong>Order ID:</strong> <span class="font-semibold text-base "><%= order.id %></span> |
-                    <strong>Status:</strong> <span class="font-semibold"><%= order.status.titleize %></span> |
-                    <strong>Total:</strong> <span class="font-semibold text-base">$<%= order.total_amount %></span> |
-                    <strong>Day:</strong> <%= order.created_at.strftime("%Y-%m-%d") %> |
-                    <strong>Hour:</strong> <%= order.created_at.strftime("%H:%M") %>
-                </span>
-                <% if current_user&.admin? %>
-                    <% unless order.status == 'completed' || order.status == 'cancelled' %>
-                        <%= form_with(model: [:admin, order], method: :patch, local: true, html: { class: "flex items-center space-x-2 mt-2 md:mt-0" }) do |form| %>
-                            <%= form.select :status, Order.statuses.keys.map { |s| [s.titleize, s] },
-                                { selected: order.status },
-                                class: "border-2 border-gray-300 rounded-md px-3 py-1 bg-white text-gray-800 focus:outline-none focus:ring-2 focus:ring-orange-500" %>
-                            <%= form.submit "Update", class: "px-4 py-1.5 bg-orange-600 text-white font-semibold rounded-md shadow-md hover:bg-orange-700 transition-colors cursor-pointer" %>
-                        <% end %>
-                    <% end %>
-                <% end %>
-            </div>
-
-            <table class="mt-4 w-full border-collapse">
-                <thead class="bg-gray-200/80">
-                    <tr>
-                        <th class="p-3 text-left text-gray-700 font-semibold rounded-tl-md">Product</th>
-                        <th class="p-3 text-center text-gray-700 font-semibold">Quantity</th>
-                        <th class="p-3 text-center text-gray-700 font-semibold rounded-tr-md">Price at Purchase</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <% order.order_items.includes(:product).each do |item| %>
-                    <tr class="border-t border-gray-200 hover:bg-gray-50/50">
-                        <td class="p-3 text-gray-800"><%= item.product.name %></td>
-                        <td class="p-3 text-center text-gray-600"><%= item.quantity %></td>
-                        <td class="p-3 text-center text-gray-800">$<%= item.price_at_purchase %></td>
-                    </tr>
-                    <% end %>
-                </tbody>
-            </table>
-        </div>
-        <% end %>
-    </div>
+        <%= render partial: "orders/user_orders", locals: { user: user, orders: orders } %>
+    <% end %>
+    <% else %>
+    <%= render partial: "orders/user_orders", locals: { user: current_user, orders: @orders } %>
     <% end %>
 </div>
