+++ Assigment3/app/views/products/_navigation.html.erb
@@ -5,19 +5,23 @@
         </li>
         <p class='point'></p>
         <li>
-            <%= link_to 'ABOUT', '#', class: 'nav-btn' %>
+            <%= link_to 'ABOUT', root_path, class: 'nav-btn' %>
         </li>
         <p class='point'></p>
         <li>
-            <%= link_to 'SERVICE', dashboard_path, class: 'nav-btn' %>
+            <% if logged_in? && current_user.admin? %>
+                <%= link_to 'SERVICE', new_admin_product_path, class: 'nav-btn' %>
+            <% else %>
+                <%= link_to 'SERVICE', root_path, class: 'nav-btn' %>
+            <% end %>
         </li>
         <p class='point'></p>
         <li>
-            <%= link_to 'BLOG', '#', class: 'nav-btn' %>
+            <%= link_to 'BLOG', root_path, class: 'nav-btn' %>
         </li>
         <p class='point'></p>
         <li>
-            <%= link_to 'CONTACTS', '#', class: 'nav-btn' %>
+            <%= link_to 'CONTACTS', root_path, class: 'nav-btn' %>
         </li>
         <% unless logged_in? %>
             <p class='point'></p>
@@ -33,6 +37,11 @@
                 <p class='point'></p>
                 <li>
                     <%= link_to 'ADMIN DASHBOARD', admin_orders_path, class: 'nav-btn' %>
+                </li>
+            <% else %>
+                <p class='point'></p>
+                <li>
+                    <%= link_to 'MY ORDERS', orders_path, class: 'nav-btn' %>
                 </li>
             <% end %>
             <p class='point'></p>
