+++ Assigment3/app/views/products/_products_cards.html.erb
@@ -10,7 +10,7 @@
   <p class="price">Price: $<%= product.price %></p>
 
   <% if show_delete_button %>
-    <%= button_to 'Delete', product_path(product), method: :delete, data: { confirm: 'Are you sure?' }, class: 'delete-button' %>
+    <%= button_to 'Delete', admin_product_path(product), method: :delete, data: { confirm: 'Are you sure?' }, class: 'delete-button' %>
   <% elsif logged_in? && show_add_to_cart_button %>
     <%= button_to 'Add to Cart', carts_path(product_id: product.id), method: :post, class: 'add-to-cart-button' %>
   <% elsif logged_in? && is_cart && item.present? %>
